--- a/src/buildstream/_workspaces.py
+++ b/src/buildstream/_workspaces.py
@@ -428,7 +428,7 @@ class Workspaces:
     # create_workspace permanent
     #
     def save_config(self):
-        assert utils._is_main_process()
+        assert utils._is_in_main_thread()
 
         config = {
             "format-version": BST_WORKSPACE_FORMAT_VERSION,