[See original merge request on GitLab](https://gitlab.com/BuildStream/buildstream/-/merge_requests/1982)
In GitLab by [[Gitlab user @BenjaminSchubert]](https://gitlab.com/BenjaminSchubert) on Jul 3, 2020, 16:30

## Description

[//]: # (Provide a description of the MR including what it resolves)

This removes the usage of processes in the scheduler to use threads instead.

This is the first step towards #1389.

This implements the changes required to run with a threaded scheduler.

## Things to test

I have tried extensively to stop/start/kill the processes to ensure that everything still behaves accordingly.

However, more eyes would be good. Please try breaking it!


## Profiles

Two profiles to show the difference between master and now:
[master.pstat](/uploads/c5b265c334b9b558e982561e73417ade/master.pstat)
[threaded.pstat](/uploads/d2b14c8af7b04c65c6963b9bf14e1fd8/threaded.pstat)

## Benchmarks

- Finished in: 57 mins, 11 secs                                                                                        
- Number of builders: [4, 8, 12]                                                                                       
- Target files: base-files/base-files.bst
- Number of runs: 3
- Number of warmups: 1
- Python versions: py38

| action                 | python_version   | commit                                  |   median time (secs) | mean time (secs) ± std   |
|:-----------------------|:-----------------|:----------------------------------------|---------------------:|:-------------------------|
| build - 12             | py38             | bschubert/no-multiprocessing - 6b2941a6 |                87.24 | 87.30 ± 0.28             |
|                        | py38             | master - 718e60f7                       |               124.65 | 124.55 ± 1.05            |
| build - 4              | py38             | bschubert/no-multiprocessing - 6b2941a6 |                89    | 88.87 ± 0.29             |
|                        | py38             | master - 718e60f7                       |               126.05 | 125.80 ± 0.77            |
| build - 8              | py38             | bschubert/no-multiprocessing - 6b2941a6 |                87.39 | 87.50 ± 0.20             |
|                        | py38             | master - 718e60f7                       |               124.23 | 123.95 ± 0.54            |
| show                   | py38             | bschubert/no-multiprocessing - 6b2941a6 |                 4.58 | 4.59 ± 0.01              |
|                        | py38             | master - 718e60f7                       |                 4.58 | 4.53 ± 0.09              |
| show - cached          | py38             | bschubert/no-multiprocessing - 6b2941a6 |                 8.99 | 8.96 ± 0.09              |
|                        | py38             | master - 718e60f7                       |                 8.9  | 8.90 ± 0.08              |
| show - sources fetched | py38             | bschubert/no-multiprocessing - 6b2941a6 |                 6.24 | 6.25 ± 0.05              |
|                        | py38             | master - 718e60f7                       |                 6.56 | 6.54 ± 0.29              |
| source fetch           | py38             | bschubert/no-multiprocessing - 6b2941a6 |                40.26 | 40.50 ± 0.43             |
|                        | py38             | master - 718e60f7                       |                61.54 | 61.78 ± 0.57             |

| action                 | python_version   | commit                                  |   median max memory (MB) | mean max memory (MB) ± std   |
|:-----------------------|:-----------------|:----------------------------------------|-------------------------:|:-----------------------------|
| build - 12             | py38             | bschubert/no-multiprocessing - 6b2941a6 |                  252.418 | 252.23 ± 0.46                |
|                        | py38             | master - 718e60f7                       |                  206.629 | 206.60 ± 0.18                |
| build - 4              | py38             | bschubert/no-multiprocessing - 6b2941a6 |                  245.957 | 245.86 ± 0.21                |
|                        | py38             | master - 718e60f7                       |                  206.613 | 206.64 ± 0.17                |
| build - 8              | py38             | bschubert/no-multiprocessing - 6b2941a6 |                  249.613 | 249.79 ± 0.31                |
|                        | py38             | master - 718e60f7                       |                  206.488 | 206.46 ± 0.11                |
| show                   | py38             | bschubert/no-multiprocessing - 6b2941a6 |                  183.023 | 183.03 ± 0.08                |
|                        | py38             | master - 718e60f7                       |                  183.273 | 183.26 ± 0.08                |
| show - cached          | py38             | bschubert/no-multiprocessing - 6b2941a6 |                  208.59  | 208.59 ± 0.11                |
|                        | py38             | master - 718e60f7                       |                  209.129 | 209.12 ± 0.04                |
| show - sources fetched | py38             | bschubert/no-multiprocessing - 6b2941a6 |                  181.305 | 181.32 ± 0.05                |
|                        | py38             | master - 718e60f7                       |                  181.711 | 181.71 ± 0.02                |
| source fetch           | py38             | bschubert/no-multiprocessing - 6b2941a6 |                  187.656 | 187.93 ± 0.68                |
|                        | py38             | master - 718e60f7                       |                  184.359 | 184.36 ± 0.07                |

There were 3 runs of each command.