@Cynical-Optimist 2020-10-13
In GitLab by [[Gitlab user @BenjaminSchubert]](https://gitlab.com/BenjaminSchubert) on Jul 4, 2020, 13:14

added 1 commit

<ul><li>a4133491 - _messenger.py: Make `timed_suspendable` public and use it in job.py</li></ul>

[Compare with previous version](/BuildStream/buildstream/-/merge_requests/1982/diffs?diff_id=99199501&start_sha=2aad85dc82d532afef179bafde12ba679abacb4c)