@Cynical-Optimist 2020-10-13
In GitLab by [[Gitlab user @BenjaminSchubert]](https://gitlab.com/BenjaminSchubert) on Jul 4, 2020, 15:09

added 2 commits

<ul><li>5c00871d - scheduler.py: Use threads instead of processes for jobs</li><li>33641df8 - _messenger.py: Make `timed_suspendable` public and use it in job.py</li></ul>

[Compare with previous version](/BuildStream/buildstream/-/merge_requests/1982/diffs?diff_id=99204343&start_sha=a413349145d8824973aaeaab4056556271447365)