@Cynical-Optimist 2020-10-13
In GitLab by [[Gitlab user @BenjaminSchubert]](https://gitlab.com/BenjaminSchubert) on Jul 5, 2020, 17:37

added 2 commits

<ul><li>a30df4b6 - scheduler.py: Use threads instead of processes for jobs</li><li>ccfca4f5 - _messenger.py: Make `timed_suspendable` public and use it in job.py</li></ul>

[Compare with previous version](/BuildStream/buildstream/-/merge_requests/1982/diffs?diff_id=99262045&start_sha=7ca6e6017cdf782ce729b0c7ddea3cd354cd0b0d)