@Cynical-Optimist 2020-10-13
In GitLab by [[Gitlab user @BenjaminSchubert]](https://gitlab.com/BenjaminSchubert) on Jul 6, 2020, 19:52

added 1 commit

<ul><li>0f840ace - job.py: Remove &#39;SUBCLASS_CUSTOM_MESSAGE&#39;, it is never used</li></ul>

[Compare with previous version](/BuildStream/buildstream/-/merge_requests/1982/diffs?diff_id=99445173&start_sha=dd3465ff6eff31920d2f50318e7df9a36b378e35)