@Cynical-Optimist 2020-10-13
In GitLab by [[Gitlab user @BenjaminSchubert]](https://gitlab.com/BenjaminSchubert) on Jul 7, 2020, 18:30

added 16 commits

<ul><li>0f840ace...99d827fa - 10 commits from branch <code>master</code></li><li>090b0271 - _signals.py: allow calling signal handler from non-main threads</li><li>4c03165b - _fixtures.py: Only get the normal number of threads at the start of session</li><li>ad9073aa - scheduler.py: Use threads instead of processes for jobs</li><li>31ec0cf7 - _messenger.py: Make `timed_suspendable` public and use it in job.py</li><li>ee93c582 - job.py: Remove &#39;SUBCLASS_CUSTOM_MESSAGE&#39;, it is never used</li><li>aaf2fcda - job.py: Remove the ability to send back child data</li></ul>

[Compare with previous version](/BuildStream/buildstream/-/merge_requests/1982/diffs?diff_id=99660539&start_sha=0f840acee9fa85bafc4eca2e28a39c182b3517b6)