@Cynical-Optimist 2020-10-13
In GitLab by [[Gitlab user @BenjaminSchubert]](https://gitlab.com/BenjaminSchubert) on Jul 7, 2020, 20:39

added 2 commits

<ul><li>0904e74e - _messenger.py: Add type annotations on the module to help refactors</li><li>15c2c44a - _messenger.py: Extract job-specific actions into another class</li></ul>

[Compare with previous version](/BuildStream/buildstream/-/merge_requests/1982/diffs?diff_id=99680662&start_sha=aaf2fcda89686f58e01e94096515890a3f077664)