@Cynical-Optimist 2020-10-13
In GitLab by [[Gitlab user @BenjaminSchubert]](https://gitlab.com/BenjaminSchubert) on Jul 7, 2020, 22:38

added 2 commits

<ul><li>6ea36b0f - _messenger.py: Extract job-specific actions into another class</li><li>82197ded - _messenger.py: Introduce a &#39;root&#39; message handler</li></ul>

[Compare with previous version](/BuildStream/buildstream/-/merge_requests/1982/diffs?diff_id=99696833&start_sha=15c2c44a8b335077773d663bf5be354d28e3d77e)