@Cynical-Optimist 2020-10-13
In GitLab by [[Gitlab user @BenjaminSchubert]](https://gitlab.com/BenjaminSchubert) on Jul 8, 2020, 20:17

added 1 commit

<ul><li>ce5b818d - _messenger.py: Make the messenger aware of jobs and stop having multiple</li></ul>

[Compare with previous version](/BuildStream/buildstream/-/merge_requests/1982/diffs?diff_id=99904385&start_sha=f26ab9fdbbfc297a7ddea7a7b2312573273f4b73)